pyproject.toml
[tool.poetry]
name = "whitesnake-projects"
version = "0.1.0"
description = ""
authors = ["Tamagil <tamagiltamagil@gmail.com>"]
readme = "README.md"
packages = [{ include = "todo" }]

[tool.poetry.dependencies]
python = "^3.11"
django = "5.1.1"
pre-commit = "3.8.0"
djangorestframework = "3.15.2"
python-dotenv = "1.0.1"
celery = {version = "5.2.7", extras = ["redis"]}
psycopg2 = "^2.9.10"
psycopg2-binary = "2.9.9"
amqp = ">=5.0.9"
asgiref = "3.8.1"
async-timeout = "4.0.3"
billiard = "3.6.4.0"
cfgv = "3.4.0"
distlib = "0.3.8"
filelock = "3.16.1"
identify = "2.6.1"
kombu = "5.2.3"
nodeenv = "1.9.1"
platformdirs = "4.3.6"
pytz = "2024.2"
pyyaml = "6.0.2"
redis = "5.1.0"
sqlparse = "0.5.1"
tzdata = "2024.2"
vine = ">=5.0.0"
virtualenv = "20.26.6"
django-rest-framework = "^0.1.0"
django-filter = "^24.3"
django-extensions = "^3.2.3"
logging = "^0.4.9.6"
django-celery-beat = "^2.7.0"
django-redis = "^5.4.0"
django-ses = "^4.2.0"
aioboto3 = "^13.2.0"
drf-yasg = "^1.21.8"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
